feat: Add "Notify Selected Entrants" button
Adds a "Notify Selected Entrants" button to the `org_my_event_details_activity.xml` layout. This button is intended to trigger a notification process for event winners.

In `OrgMyEventDetailsActivity`, the logic for updating participant statuses to "SELECTED" and setting the `chosen_entrants` field after drawing winners has been removed in preparation for the new notification functionality.